refactor: GitHub Copilotのレビューに基づいてテストコードを改善
Copilotのセキュリティとパフォーマンスに関する指摘に対応：

1. テスト設定の共通化
   - utils/testConfig.cjs を作成
   - headless, maxWorkers, enableScreenshots を環境変数で制御可能に
   - 全テストファイルで共通設定を使用

2. セキュリティの改善
   - Workerコードを utils/testWorker.js に分離
   - eval: true の使用を排除（セキュリティリスクの解消）

3. パフォーマンスの最適化
   - スクリーンショット機能を環境変数で制御（ENABLE_SCREENSHOTS）
   - ワーカー数を環境変数で調整可能（MAX_WORKERS）

4. ドキュメントの更新
   - 環境変数の使用方法を追加
   - 改善内容を明記

これにより、テストの保守性、セキュリティ、パフォーマンスが向上しました。
