+++ tests/e2e/test-stage0-4-simple.cjs
@@ -1,8 +1,9 @@
 const GameTestHelpers = require('./utils/GameTestHelpers.cjs');
+const testConfig = require('./utils/testConfig.cjs');
 
 async function runTest() {
     const test = new GameTestHelpers({ 
-        headless: false,  // ブラウザを表示
+        headless: testConfig.headless,
         verbose: true,
         timeout: 60000  // 60秒のタイムアウト
     });
